backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# 机器学习
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
shap==0.43.0
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# 深度学习（医学影像识别）- 使用 CPU 版本
# 注意：使用以下命令单独安装 PyTorch CPU 版本（更小，无需 CUDA）
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu

# 图像处理
pillow>=9.0.0
opencv-python>=4.5.0
scipy>=1.9.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 进度条
tqdm>=4.62.0

# 工具
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1
requests>=2.28.0

# 阿里云OSS
oss2>=2.18.0

# 大模型
zhipuai>=2.1.5  # 智谱AI

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
black==23.11.0
flake8==6.1.0

